Use tokio::fs::copy for copying file data
By using `tokio::fs::copy` we are able to use the `copy_file_range` system call in Linux which dramatically speeds up the copies (at least 5-20x in our tests).

Note the `tokio::io::copy(_buf)` funcitons don't seem to be able to use this syscall reliably in our testing.

Unfortunately, the `tokio::fs::copy` call is path based and that means potentially having to read the destination metadata twice. For this reason, we add `is_fresh` variable to avoid having to check the destination metadata whenever we're sure it doesn't exist.